refactoring estyle
